+++ src/main/kotlin/io/acari/doki/notification/UpdateNotification.kt
@@ -8,6 +8,7 @@ val UPDATE_MESSAGE: String = """
       <ul>
       <li>Doki Themed Material Icons, see settings to enable them.</li>
       <li>Brought back configurable theme name in Status Bar.</li>
+      <li>Updated Light Natsuki's color scheme.</li>
       </ul>
       <br>Please see the <a href="https://github.com/cyclic-reference/ddlc-jetbrains-theme/blob/master/changelog/CHANGELOG.md">Changelog</a> for more details.
       <br>