+++ changelog/CHANGELOG.md
@@ -7,7 +7,7 @@
         - File Icons
         - PSI Icons
     - Icons are off by default, so be sure to go to the theme settings to turn them on.
-
+- Updated Light Natsuki's color scheme. 
 # 5.0.3
 
 - Fixed Stickers not showing up in certain environments.